Moved to more obvious MVC framework
Changed the "resource" idea to a "document" idea, because resource
implies something like a picture or something that is not modified.
RobotModel switched to CommonLib, because it contains more than just the
model.  There will be two main views for the model; a view for the
editor, and a view for the robot.  The controller is yet to be decided