Initial work on designer view + reorganization
Started working on the OzRobotBuilder.NET project for actual design.
Switched over to a single-document setting to reduce complexity (scripts
might be integrated into the json, or loaded as separate documents
referenced in the json).  Tested document saving system (works properly)
for files.  Loading is not tested.  Recreated the CommonLib project so
it is called "CommonLib" from the start instead of it being "RobotModel"
renamed to "CommonLib."  Finally, started using a library
"JSonPrettyPrinter" to improve output format of the saved files to make
direct editing of them possible before the fancy editor is done.