Removed old classes, updated docs, and finished implementation
Removed the first version of the validation which was a second set of
classes along side each data type.  Also updated the documents for all
of the validation so it is complete.  Finally, the implementation for
all of the validation for all of the data types is usably complete.  It
does not include some kinds of validation, but those are mostly
warning-type validations