First .NET core/MVVM/WPF/VSIX commit
wow this was poorly organized, but its a lot of work i don't want to
lose.  Changed to .NET Core, the MVVM patter, and began work on WPF
editor for a visual studio plugin.  This project's final form is
changing from a standalone app to a visual studio extension, however all
of the non-visual-studio-specific features will make their way into .NET
core libraries and potentially command line tools.  The WPF controls and
MVVM pattern not specific to visual studio will be packaged away in a
.NET 4.6.1 assembly so non-visual studio use is possible without copying
any code.  The current code base is mostly dysfunctional, however the
model is ok (despite not implementing the validation interface from the
component model.)