Removed All Choices in XSD Schema
since these were creating "object" arrays in the generated c# code, I
wanted to get rid of them.  The Potentiometer was removed, because it
can use the AnalogInput type.  The choice for digital inputs and quad
encoders was unncessessary.  ID reuse would become a problem before too
many items would, therefore the max number is restricted to each of them
to allow all of the digital inputs on the roborio to go to either
option.