Made ClampValues property part of IOService
This choice was made, because this value will be global, and is tightly
tied with the IO service itself, despite not being a hardware-access
feature.  It makes the most sense to put it here, because the ViewModel,
which is responsible for the clamping choice, now has access to this
configuration value.