Removed DataAccess service
This was a bad solution to the problem at hand.  Since the visual studio
plugin only has one instance and multiple instances of the view-model,
this would not work responsibly, therefore the "TopParent," or rather
the robot instance, was added to the "Provider" interface.  Now the
view-model structure is more modular.