Most of first draft done
The socket system was given a client-server aspect.  There is also a
header in the request of 4 bytes which contains the length of the
request.  Network streams are now being used as well.  Many of the
methods were moved to the Utils class so that everyone can have access
to these request sending/receiving as long as they have a network
stream.