Refocused Commit
After working on this project for an extended period of time with only a
small duration each time resulted in duplicate work done worse the
second time.  The existing code for two-way communication is used and
the old attempt is deleted.  The multi-connection system is completely
removed, becuase the current state of it was terrible.  There is still
an opportunity of having the `CompoundRequestDelegator` through the
`Body` part of the request.